Use short to store block/biome id

+++ src/server/world/palette.cpp
@@ -1,22 +1,21 @@
 #include "palette.h"
 
 #include <algorithm>
-#include <bitset>
 #include <iostream>
 #include <utility>
 
-Palette::Palette(int bpe, std::set<int> palette)
+Palette::Palette(int bpe, std::set<short> palette)
     : bitsPerEntry(bpe),
       palette(std::move(palette)) {}
 
 Palette::~Palette() = default;
 
-SingleValuedPalette::SingleValuedPalette(std::set<int> palette)
+SingleValuedPalette::SingleValuedPalette(std::set<short> palette)
     : Palette(0, std::move(palette)) {}
 
 SingleValuedPalette::~SingleValuedPalette() = default;
 
-std::vector<unsigned long long> SingleValuedPalette::apply(std::vector<int>) {
+std::vector<unsigned long long> SingleValuedPalette::apply(std::vector<short>) {
   return {};
 }
 
@@ -24,12 +23,12 @@ void SingleValuedPalette::write(ByteBuffer& out) {
   out.writeVarInt(*palette.begin());
 }
 
-IndirectPalette::IndirectPalette(int bpe, std::set<int> palette)
+IndirectPalette::IndirectPalette(int bpe, std::set<short> palette)
     : Palette(bpe, std::move(palette)) {}
 
 IndirectPalette::~IndirectPalette() = default;
 
-std::vector<unsigned long long> IndirectPalette::apply(std::vector<int> data) {
+std::vector<unsigned long long> IndirectPalette::apply(std::vector<short> data) {
   std::vector<unsigned long long> dst;
   unsigned long long current = 0;
   int currentBit = 0;
@@ -59,7 +58,7 @@ std::vector<unsigned long long> IndirectPalette::apply(std::vector<int> data) {
 
 void IndirectPalette::write(ByteBuffer& out) {
   out.writeVarInt((int) palette.size());
-  for (int element: palette) {
+  for (short element: palette) {
     out.writeVarInt(element);
   }
 }
@@ -69,12 +68,12 @@ DirectPalette::DirectPalette()
 
 DirectPalette::~DirectPalette() = default;
 
-std::vector<unsigned long long> DirectPalette::apply(std::vector<int> data) {
+std::vector<unsigned long long> DirectPalette::apply(std::vector<short> data) {
   std::vector<unsigned long long> dst;
   unsigned long long current = 0;
   int currentBit = 0;
 
-  for (int element: data) {
+  for (short element: data) {
     current |= ((unsigned long long) element << currentBit);
     currentBit += bitsPerEntry;
     if (currentBit >= 64) {
