+++ src/server/world/palette.h
@@ -7,35 +7,35 @@
 
 class Palette {
  public:
-  Palette(int, std::set<int>);
+  Palette(int, std::set<short>);
   virtual ~Palette();
 
  public:
-  virtual std::vector<unsigned long long> apply(std::vector<int>) = 0;
+  virtual std::vector<unsigned long long> apply(std::vector<short>) = 0;
   virtual void write(ByteBuffer&) = 0;
 
  public:
   int bitsPerEntry;
-  std::set<int> palette;
+  std::set<short> palette;
 };
 
 class SingleValuedPalette : public Palette {
  public:
-  explicit SingleValuedPalette(std::set<int>);
+  explicit SingleValuedPalette(std::set<short>);
   ~SingleValuedPalette() override;
 
  public:
-  std::vector<unsigned long long> apply(std::vector<int>) override;
+  std::vector<unsigned long long> apply(std::vector<short>) override;
   void write(ByteBuffer&) override;
 };
 
 class IndirectPalette : public Palette {
  public:
-  IndirectPalette(int, std::set<int>);
+  IndirectPalette(int, std::set<short>);
   ~IndirectPalette() override;
 
  public:
-  std::vector<unsigned long long> apply(std::vector<int>) override;
+  std::vector<unsigned long long> apply(std::vector<short>) override;
   void write(ByteBuffer&) override;
 };
 
@@ -45,6 +45,6 @@ class DirectPalette : public Palette {
   ~DirectPalette() override;
 
  public:
-  std::vector<unsigned long long> apply(std::vector<int>) override;
+  std::vector<unsigned long long> apply(std::vector<short>) override;
   void write(ByteBuffer&) override;
 };
